+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Add "Owl assistant" easter egg: a floating owl buddy in the bottom-left corner that shares productivity tips and programming jokes
+
 ## [1.12.3]
 
 ### Changed